Auth API routes (Issue #7, #8)
