+++ backend/src/routes/auth.js
@@ -1,4 +1,63 @@
 const express = require('express');
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const { PrismaClient } = require('@prisma/client');
+const { authenticateToken } = require('../middleware/auth');
+
 const router = express.Router();
-// Week 2: auth routes added here
+const prisma = new PrismaClient();
+
+// REGISTER
+router.post('/register', async (req, res) => {
+  const { name, email, password } = req.body;
+
+  if (!name || !email || !password)
+    return res.status(400).json({ error: 'All fields are required' });
+
+  const existing = await prisma.user.findUnique({ where: { email } });
+  if (existing)
+    return res.status(400).json({ error: 'Email already registered' });
+
+  const hashedPassword = await bcrypt.hash(password, 10);
+  const user = await prisma.user.create({
+    data: { name, email, password: hashedPassword }
+  });
+
+  const token = jwt.sign({ userId: user.id }, process.env.JWT_SECRET, { expiresIn: '7d' });
+
+  res.status(201).json({
+    token,
+    user: { id: user.id, name: user.name, email: user.email, role: user.role }
+  });
+});
+
+// LOGIN
+router.post('/login', async (req, res) => {
+  const { email, password } = req.body;
+
+  if (!email || !password)
+    return res.status(400).json({ error: 'Email and password required' });
+
+  const user = await prisma.user.findUnique({ where: { email } });
+  if (!user)
+    return res.status(401).json({ error: 'Invalid email or password' });
+
+  const valid = await bcrypt.compare(password, user.password);
+  if (!valid)
+    return res.status(401).json({ error: 'Invalid email or password' });
+
+  const token = jwt.sign({ userId: user.id }, process.env.JWT_SECRET, { expiresIn: '7d' });
+
+  res.json({
+    token,
+    user: { id: user.id, name: user.name, email: user.email, role: user.role }
+  });
+});
+
+// GET CURRENT USER
+router.get('/me', authenticateToken, (req, res) => {
+  const { id, name, email, role } = req.user;
+  res.json({ id, name, email, role });
+});
+
 module.exports = router;
